+++ lib/onetime/application/authorization_policies.rb
@@ -172,6 +172,86 @@ module Onetime
         true
       end
 
+      # Verify current user owns the organization
+      #
+      # Colonels (site admins) have automatic superuser bypass.
+      # Otherwise, user must be organization owner.
+      #
+      # @param organization [Onetime::Organization]
+      # @param error_key [String] I18n key for the rejection message
+      # @raise [Onetime::Forbidden] If user is not owner and not colonel
+      def verify_organization_owner(organization, error_key: 'api.organizations.errors.owner_required')
+        verify_one_of_roles!(
+          colonel: true,
+          custom_check: -> { organization.owner?(cust) },
+          error_message: I18n.t(error_key, locale: locale, default: 'Only organization owner can perform this action'),
+        )
+      end
+
+      # Verify current user is an organization member
+      #
+      # Colonels (site admins) have automatic superuser bypass.
+      # Otherwise, user must be organization member.
+      #
+      # @param organization [Onetime::Organization]
+      # @param error_key [String] I18n key for the rejection message
+      # @raise [Onetime::Forbidden] If user is not a member and not colonel
+      def verify_organization_member(organization, error_key: 'api.organizations.errors.member_required')
+        verify_one_of_roles!(
+          colonel: true,
+          custom_check: -> { organization.member?(cust) },
+          error_message: I18n.t(error_key, locale: locale, default: 'You must be an organization member to perform this action'),
+        )
+      end
+
+      # Verify current user has admin privileges in the organization
+      #
+      # Owner is implicitly admin. Colonels (site admins) bypass.
+      #
+      # @param organization [Onetime::Organization]
+      # @param error_key [String] I18n key for the rejection message; defaults to the generic
+      #   admin-required message. Per-action callers should pass their own key
+      #   (e.g. 'api.domains.errors.add_admin_required').
+      # @raise [Onetime::Forbidden] If user is not owner/admin
+      def verify_organization_admin(organization, error_key: 'api.organizations.errors.admin_required')
+        verify_one_of_roles!(
+          colonel: true,
+          custom_check: -> { organization.owner?(cust) || organization_admin?(organization) },
+          error_message: I18n.t(error_key, locale: locale, default: 'Only organization owners and admins can perform this action'),
+        )
+      end
+
+      # Check if current user holds the admin role in the organization
+      #
+      # Does not include owner — callers that want owner-or-admin should use
+      # verify_organization_admin (which composes owner? || organization_admin?).
+      #
+      # @param organization [Onetime::Organization]
+      # @return [Boolean]
+      def organization_admin?(organization)
+        return false if cust.nil?
+
+        membership = Onetime::OrganizationMembership.find_by_org_customer(
+          organization.objid, cust.objid
+        )
+        membership&.admin?
+      end
+
+      # Load organization by external ID, raising not-found on miss
+      #
+      # @param extid [String] Organization external ID
+      # @raise [Onetime::RecordNotFound] If organization not found
+      # @return [Onetime::Organization]
+      def load_organization(extid)
+        organization = Onetime::Organization.find_by_extid(extid)
+        if organization.nil?
+          raise_not_found(
+            I18n.t('api.organizations.errors.not_found', locale: locale, extid: extid, default: "Organization not found: #{extid}"),
+          )
+        end
+        organization
+      end
+
       private
 
       # Build user-friendly error message from authorization requirements
